# server/out/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/core/system/uploadImage": {
      "post": {
        "operationId": "CoreUploadController_uploadImage",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "上传图片",
        "tags": [
          "文件上传"
        ]
      }
    },
    "/api/core/system/uploadFile": {
      "post": {
        "operationId": "CoreUploadController_uploadFile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "上传文件",
        "tags": [
          "文件上传"
        ]
      }
    },
    "/api/core/system/chunkUpload": {
      "post": {
        "operationId": "CoreUploadController_chunkUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkFileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "分片上传",
        "tags": [
          "文件上传"
        ]
      }
    },
    "/api/system/attachment/list": {
      "get": {
        "operationId": "AttachmentController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "附件列表",
        "tags": [
          "附件管理"
        ]
      }
    },
    "/api/system/attachment/detail/{id}": {
      "get": {
        "operationId": "AttachmentController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "附件详情",
        "tags": [
          "附件管理"
        ]
      }
    },
    "/api/system/attachment/upload": {
      "post": {
        "operationId": "AttachmentController_upload",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "上传附件",
        "tags": [
          "附件管理"
        ]
      }
    },
    "/api/system/attachment/update/{id}": {
      "put": {
        "operationId": "AttachmentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "更新附件",
        "tags": [
          "附件管理"
        ]
      }
    },
    "/api/system/attachment/delete/{id}": {
      "delete": {
        "operationId": "AttachmentController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "删除附件",
        "tags": [
          "附件管理"
        ]
      }
    },
    "/api/system/attachment/batchDelete": {
      "delete": {
        "operationId": "AttachmentController_batchDelete",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "批量删除附件",
        "tags": [
          "附件管理"
        ]
      }
    },
    "/api/system/attachment/move": {
      "put": {
        "operationId": "AttachmentController_move",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "移动附件",
        "tags": [
          "附件管理"
        ]
      }
    },
    "/api/system/attachment/download/{id}": {
      "get": {
        "operationId": "AttachmentController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "下载附件",
        "tags": [
          "附件管理"
        ]
      }
    },
    "/api/system/attachment/stats": {
      "get": {
        "operationId": "AttachmentController_stats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "附件统计",
        "tags": [
          "附件管理"
        ]
      }
    },
    "/api/system/attachment-category/list": {
      "get": {
        "operationId": "AttachmentCategoryController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "分类列表",
        "tags": [
          "附件分类"
        ]
      }
    },
    "/api/system/attachment-category/detail/{id}": {
      "get": {
        "operationId": "AttachmentCategoryController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "分类详情",
        "tags": [
          "附件分类"
        ]
      }
    },
    "/api/system/attachment-category/create": {
      "post": {
        "operationId": "AttachmentCategoryController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "创建分类",
        "tags": [
          "附件分类"
        ]
      }
    },
    "/api/system/attachment-category/update/{id}": {
      "put": {
        "operationId": "AttachmentCategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "更新分类",
        "tags": [
          "附件分类"
        ]
      }
    },
    "/api/system/attachment-category/delete/{id}": {
      "delete": {
        "operationId": "AttachmentCategoryController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "删除分类",
        "tags": [
          "附件分类"
        ]
      }
    },
...
  },
  "info": {
    "title": "FssAdmin",
    "description": "FssAdmin API 文档",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "ChunkFileDto": {
        "type": "object",
        "properties": {
          "index": {
            "type": "number"
          },
          "totalChunks": {
            "type": "number"
          },
          "uploadId": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          }
        },
        "required": [
          "index",
          "totalChunks",
          "uploadId",
          "fileName"
        ]
      },
...
    }
  }
}